backend/requirements.txt.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
python-dateutil